chore(docker): Fixed issue with proxies not showing right ip
This patch fixes an issue when running nginx behind a reverse proxy. The reverse proxy lists its own ip as the "remote_addr" instead of the user's actual ip. Updating nginx to use "http_x_forwarded_for" instead of "remote_addr", along with configuring the reverse proxy to pass the x-forwarded-for header fixes this issue.